+++ sdk/python/packages/flet-ads/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Refactored `BaseAd` class to be based on `flet.BaseControl`.
+- Refactored `InterstitialAd` to a service-based control (`flet.Service`).
+- Refactored `BannerAd` to a layout control (`flet.LayoutControl`).
+- Updated Example app.
+
 ## 0.80.0
 
 ### Added